fulfillment/pyproject.toml
[project]
name = "grooved-fulfillment"
version = "0.1.0"
description = "Fulfillment module for Grooved Learning - handles shipping labels and packing slips"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "reportlab>=4.0.0",
    "backoff>=2.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]